campus_lost_and_found/requirements.txt
# 基础框架
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# 用户认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 环境变量
python-dotenv==1.0.0

# AI 模型
openai==1.6.1

# 图片处理
pillow==10.1.0

# 文本相似度
scikit-learn==1.3.2
jieba==0.42.1

# HTTP 客户端
httpx==0.25.2
